Create assert method - mostly for param validation
Fixes #106
